src/agents-api/requirements.txt
google-generativeai>=0.3.0
Flask>=2.0.0
python-dotenv>=0.19.0
Flask-Cors
